requirements.txt
requests>=2.28.0
GitPython>=3.1.30
python-dotenv>=1.0.0
